Vid2Vid: Part III

+++ ui-src/Components/VideoFrame.cs
@@ -6,6 +6,12 @@ using Windows.Devices.Geolocation;
 
 namespace SD_FXUI
 {
+    struct LoadData
+    {
+        public int FramePerSecond;
+        public string File;
+    }
+
     public struct VideoData
     {
         public bool FrameDone;
@@ -16,6 +22,8 @@ namespace SD_FXUI
 
     internal class VideoFrame
     {
+        static LoadData LastVideoLoad;
+
         static public string CapturePath = FS.GetImagesDir() + "vidcap\\";
         static void SaveFrames(BitmapSource Src, uint It)
         {
@@ -31,39 +39,64 @@ namespace SD_FXUI
                 encoder.Save(fileStream);
             }
         }
-        public static void ReadVideo(string Path)
+        public static void ReadVideo(string Path, int FPS)
         {
-            FS.Dir.Delete(CapturePath, true);
-            Directory.CreateDirectory(CapturePath);
+            bool bCheck = LastVideoLoad.File == Path && LastVideoLoad.FramePerSecond == FPS;
 
-            GlobalVariables.LastVideoData.FrameDone = false;
-            GlobalVariables.LastVideoData.ActiveRender = false;
-            GlobalVariables.LastVideoData.ActiveFrame = 0;
-            GlobalVariables.LastVideoData.Files = new List<string>();
-
-            VideoCapture Capture = new VideoCapture(Path);
-            var image = new Mat();
-
-            uint FrameIt = 0;
-
-            while (Capture.IsOpened())
+            if (!bCheck)
             {
-                Capture.Read(image);
+                Host.Print("Extract video captures...");
 
-                if (image.Empty())
-                    break;
+                FS.Dir.Delete(CapturePath, true);
+                Directory.CreateDirectory(CapturePath);
 
-                SaveFrames(OpenCvSharp.WpfExtensions.BitmapSourceConverter.ToBitmapSource(image), FrameIt);
+                GlobalVariables.LastVideoData.FrameDone = false;
+                GlobalVariables.LastVideoData.ActiveRender = true;
+                GlobalVariables.LastVideoData.ActiveFrame = 0;
+                GlobalVariables.LastVideoData.Files = new List<string>();
 
-                FrameIt++;
+                VideoCapture Capture = new VideoCapture(Path);
+                var ImgMat = new Mat();
 
-                if (Cv2.WaitKey(1) == 113)
-                    break;
+                int DeltaFPS = 0;
+                if (FPS < Capture.Fps)
+                {
+                    double fDeltaFPS = (double)FPS / Capture.Fps;
+                    fDeltaFPS *= 10;
+                    DeltaFPS = (int)fDeltaFPS;
+                }
+                else
+                {
+                    DeltaFPS = 999999999;
+                }
+
+                uint FrameIt = 0;
+
+                while (Capture.IsOpened())
+                {
+                    Capture.Read(ImgMat);
+
+                    if (ImgMat.Empty())
+                        break;
+
+                    if (FrameIt % DeltaFPS != 0)
+                    {
+                        SaveFrames(OpenCvSharp.WpfExtensions.BitmapSourceConverter.ToBitmapSource(ImgMat), FrameIt);
+                    }
+
+                    FrameIt++;
+
+                    if (Cv2.WaitKey(1) == 113)
+                        break;
+                }
+
+                LastVideoLoad.File = Path;
+                LastVideoLoad.FramePerSecond = FPS;
             }
-            //RenderVideo();
         }
         public static void RenderVideo(string Input)
         {
+            Host.Print("Render video...");
             VideoCapture InCapture = new VideoCapture($"{Input}%d.png");
             //VideoCapture OutCapture = new VideoCapture("N:\\TestMP4\\test.mp4", VideoCaptureAPIs.FFMPEG);
